:construction: CameraManager: Dissolved `class CameraBehaviorOrbit`
Intermediate change - functions were attached as `static` to CameraManager.
