+++ source/main/gfx/camera/CameraManager.cpp
@@ -23,6 +23,7 @@
 #include "Application.h"
 #include "BeamFactory.h"
 #include "Character.h"
+#include "Collisions.h"
 #include "DepthOfFieldEffect.h"
 #include "GUI_GameConsole.h"
 #include "InputEngine.h"
@@ -33,7 +34,6 @@
 #include "GUIManager.h"
 #include "PerVehicleCameraContext.h"
 
-#include "CameraBehaviorOrbit.h"
 #include "CameraBehaviorCharacter.h"
 #include "CameraBehaviorFree.h"
 #include "CameraBehaviorVehicle.h"
@@ -323,12 +323,14 @@ void CameraManager::DeactivateCurrentBehavior()
 {
     switch(m_current_behavior)
     {
-    case CAMERA_BEHAVIOR_CHARACTER:       m_cam_behav_character      ->deactivate(ctx);  return;
+    case CAMERA_BEHAVIOR_CHARACTER:       return;
+
     case CAMERA_BEHAVIOR_STATIC:
         gEnv->mainCamera->setFOVy(m_staticcam_previous_fov);
         return;
-    case CAMERA_BEHAVIOR_VEHICLE:         m_cam_behav_vehicle        ->deactivate(ctx);  return;
-    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  m_cam_behav_vehicle_spline ->deactivate(ctx);  return;
+
+    case CAMERA_BEHAVIOR_VEHICLE:         return;
+    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  return;
     case CAMERA_BEHAVIOR_VEHICLE_CINECAM: m_cam_behav_vehicle_cinecam->deactivate(ctx);  return;
     case CAMERA_BEHAVIOR_FREE:            m_cam_behav_free           ->deactivate(ctx);  return;
     case CAMERA_BEHAVIOR_FIXED:           return;
@@ -400,9 +402,9 @@ bool CameraManager::mouseMoved(const OIS::MouseEvent& _arg)
     {
     case CAMERA_BEHAVIOR_CHARACTER:       return m_cam_behav_character      ->mouseMoved(ctx, _arg);
     case CAMERA_BEHAVIOR_STATIC:          return false;
-    case CAMERA_BEHAVIOR_VEHICLE:         return m_cam_behav_vehicle        ->mouseMoved(ctx, _arg);
+    case CAMERA_BEHAVIOR_VEHICLE:         return CameraBehaviorOrbitMouseMoved(ctx, _arg);
     case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  return m_cam_behav_vehicle_spline ->mouseMoved(ctx, _arg);
-    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: return m_cam_behav_vehicle_cinecam->mouseMoved(ctx, _arg);
+    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: return CameraBehaviorOrbitMouseMoved(ctx, _arg);
     case CAMERA_BEHAVIOR_FREE:            return m_cam_behav_free           ->mouseMoved(ctx, _arg);
     case CAMERA_BEHAVIOR_FIXED:           return false;
     case CAMERA_BEHAVIOR_ISOMETRIC:       return false;
@@ -415,7 +417,7 @@ bool CameraManager::mousePressed(const OIS::MouseEvent& _arg, OIS::MouseButtonID
 {
     switch(m_current_behavior)
     {
-    case CAMERA_BEHAVIOR_CHARACTER:       return m_cam_behav_character      ->mousePressed(ctx, _arg, _id);
+    case CAMERA_BEHAVIOR_CHARACTER:       return false;
     case CAMERA_BEHAVIOR_STATIC:          return false;
     case CAMERA_BEHAVIOR_VEHICLE:         return m_cam_behav_vehicle        ->mousePressed(ctx, _arg, _id);
     case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  return m_cam_behav_vehicle_spline ->mousePressed(ctx, _arg, _id);
@@ -432,11 +434,11 @@ bool CameraManager::mouseReleased(const OIS::MouseEvent& _arg, OIS::MouseButtonI
 {
     switch(m_current_behavior)
     {
-    case CAMERA_BEHAVIOR_CHARACTER:       return m_cam_behav_character      ->mouseReleased(ctx, _arg, _id);
+    case CAMERA_BEHAVIOR_CHARACTER:       return false;
     case CAMERA_BEHAVIOR_STATIC:          return false;
-    case CAMERA_BEHAVIOR_VEHICLE:         return m_cam_behav_vehicle        ->mouseReleased(ctx, _arg, _id);
-    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  return m_cam_behav_vehicle_spline ->mouseReleased(ctx, _arg, _id);
-    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: return m_cam_behav_vehicle_cinecam->mouseReleased(ctx, _arg, _id);
+    case CAMERA_BEHAVIOR_VEHICLE:         return false;
+    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  return false;
+    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: return false;
     case CAMERA_BEHAVIOR_FREE:            return m_cam_behav_free           ->mouseReleased(ctx, _arg, _id);
     case CAMERA_BEHAVIOR_FIXED:           return false;
     case CAMERA_BEHAVIOR_ISOMETRIC:       return false;
@@ -461,11 +463,11 @@ void CameraManager::NotifyContextChange()
 {
     switch(m_current_behavior)
     {
-    case CAMERA_BEHAVIOR_CHARACTER:       m_cam_behav_character      ->notifyContextChange(ctx);  return;
+    case CAMERA_BEHAVIOR_CHARACTER:       CameraBehaviorOrbitNotifyContextChange(ctx);  return;
     case CAMERA_BEHAVIOR_STATIC:          return;
-    case CAMERA_BEHAVIOR_VEHICLE:         m_cam_behav_vehicle        ->notifyContextChange(ctx);  return;
-    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  m_cam_behav_vehicle_spline ->notifyContextChange(ctx);  return;
-    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: m_cam_behav_vehicle_cinecam->notifyContextChange(ctx);  return;
+    case CAMERA_BEHAVIOR_VEHICLE:         CameraBehaviorOrbitNotifyContextChange(ctx);  return;
+    case CAMERA_BEHAVIOR_VEHICLE_SPLINE:  CameraBehaviorOrbitNotifyContextChange(ctx);  return;
+    case CAMERA_BEHAVIOR_VEHICLE_CINECAM: CameraBehaviorOrbitNotifyContextChange(ctx);  return;
     case CAMERA_BEHAVIOR_FREE:            m_cam_behav_free           ->notifyContextChange(ctx);  return;
     case CAMERA_BEHAVIOR_FIXED:           return;
     case CAMERA_BEHAVIOR_ISOMETRIC:       return;
@@ -626,4 +628,163 @@ void CameraManager::UpdateCameraBehaviorStatic(const CameraManager::CameraContex
     gEnv->mainCamera->setPosition(m_staticcam_position);
     gEnv->mainCamera->lookAt(lookAt);
     gEnv->mainCamera->setFOVy(Radian(fov));
+}
+
+void CameraManager::CameraBehaviorOrbitUpdate( CameraManager::CameraContext& ctx)
+{
+    if (RoR::App::GetInputEngine()->getEventBoolValueBounce(EV_CAMERA_LOOKBACK))
+    {
+        if (ctx.camRotX > Degree(0))
+        {
+            ctx.camRotX = Degree(0);
+        }
+        else
+        {
+            ctx.camRotX = Degree(180);
+        }
+    }
+
+    ctx.camRotX += (RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_ROTATE_RIGHT) - RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_ROTATE_LEFT)) * ctx.cct_rot_scale;
+    ctx.camRotY += (RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_ROTATE_UP) - RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_ROTATE_DOWN)) * ctx.cct_rot_scale;
+
+    ctx.camRotY = std::max((Radian)Degree(-80), ctx.camRotY);
+    ctx.camRotY = std::min(ctx.camRotY, (Radian)Degree(88));
+
+    ctx.camRotXSwivel = (RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_SWIVEL_RIGHT) - RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_SWIVEL_LEFT)) * Degree(90);
+    ctx.camRotYSwivel = (RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_SWIVEL_UP) - RoR::App::GetInputEngine()->getEventValue(EV_CAMERA_SWIVEL_DOWN)) * Degree(60);
+
+    ctx.camRotYSwivel = std::max((Radian)Degree(-80) - ctx.camRotY, ctx.camRotYSwivel);
+    ctx.camRotYSwivel = std::min(ctx.camRotYSwivel, (Radian)Degree(88) - ctx.camRotY);
+
+    if (RoR::App::GetInputEngine()->getEventBoolValue(EV_CAMERA_ZOOM_IN) && ctx.camDist > 1)
+    {
+        ctx.camDist -= ctx.cct_trans_scale;
+    }
+    if (RoR::App::GetInputEngine()->getEventBoolValue(EV_CAMERA_ZOOM_IN_FAST) && ctx.camDist > 1)
+    {
+        ctx.camDist -= ctx.cct_trans_scale * 10;
+    }
+    if (RoR::App::GetInputEngine()->getEventBoolValue(EV_CAMERA_ZOOM_OUT))
+    {
+        ctx.camDist += ctx.cct_trans_scale;
+    }
+    if (RoR::App::GetInputEngine()->getEventBoolValue(EV_CAMERA_ZOOM_OUT_FAST))
+    {
+        ctx.camDist += ctx.cct_trans_scale * 10;
+    }
+
+    if (RoR::App::GetInputEngine()->getEventBoolValue(EV_CAMERA_RESET))
+    {
+        CameraBehaviorOrbitReset(ctx);
+    }
+
+    if (RoR::App::GetInputEngine()->isKeyDown(OIS::KC_RSHIFT) && RoR::App::GetInputEngine()->isKeyDownValueBounce(OIS::KC_SPACE))
+    {
+        ctx.limitCamMovement = !ctx.limitCamMovement;
+        if (ctx.limitCamMovement)
+        {
+            RoR::App::GetConsole()->putMessage(Console::CONSOLE_MSGTYPE_INFO, Console::CONSOLE_SYSTEM_NOTICE, _L("Limited camera movement enabled"), "camera_go.png", 3000);
+            RoR::App::GetGuiManager()->PushNotification("Notice:", _L("Limited camera movement enabled"));
+        }
+        else
+        {
+            RoR::App::GetConsole()->putMessage(Console::CONSOLE_MSGTYPE_INFO, Console::CONSOLE_SYSTEM_NOTICE, _L("Limited camera movement disabled"), "camera_go.png", 3000);
+            RoR::App::GetGuiManager()->PushNotification("Notice:", _L("Limited camera movement disabled"));
+        }
+    }
+
+    if (ctx.limitCamMovement && ctx.camDistMin > 0.0f)
+    {
+        ctx.camDist = std::max(ctx.camDistMin, ctx.camDist);
+    }
+    if (ctx.limitCamMovement && ctx.camDistMax > 0.0f)
+    {
+        ctx.camDist = std::min(ctx.camDist, ctx.camDistMax);
+    }
+
+    ctx.camDist = std::max(0.0f, ctx.camDist);
+
+    Vector3 desiredPosition = ctx.camLookAt + ctx.camDist * 0.5f * Vector3(
+        sin(ctx.targetDirection.valueRadians() + (ctx.camRotX - ctx.camRotXSwivel).valueRadians()) * cos(ctx.targetPitch.valueRadians() + (ctx.camRotY - ctx.camRotYSwivel).valueRadians())
+        , sin(ctx.targetPitch.valueRadians() + (ctx.camRotY - ctx.camRotYSwivel).valueRadians())
+        , cos(ctx.targetDirection.valueRadians() + (ctx.camRotX - ctx.camRotXSwivel).valueRadians()) * cos(ctx.targetPitch.valueRadians() + (ctx.camRotY - ctx.camRotYSwivel).valueRadians())
+    );
+
+    if (ctx.limitCamMovement && App::GetSimTerrain())
+    {
+        float h = App::GetSimTerrain()->GetHeightAt(desiredPosition.x, desiredPosition.z) + 1.0f;
+
+        desiredPosition.y = std::max(h, desiredPosition.y);
+    }
+
+    if (ctx.camLookAtLast == Vector3::ZERO)
+    {
+        ctx.camLookAtLast = ctx.camLookAt;
+    }
+    if (ctx.camLookAtSmooth == Vector3::ZERO)
+    {
+        ctx.camLookAtSmooth = ctx.camLookAt;
+    }
+    if (ctx.camLookAtSmoothLast == Vector3::ZERO)
+    {
+        ctx.camLookAtSmoothLast = ctx.camLookAtSmooth;
+    }
+
+    Vector3 camDisplacement = ctx.camLookAt - ctx.camLookAtLast;
+    Vector3 precedingLookAt = ctx.camLookAtSmoothLast + camDisplacement;
+    Vector3 precedingPosition = gEnv->mainCamera->getPosition() + camDisplacement;
+
+    Vector3 camPosition = (1.0f / (ctx.camRatio + 1.0f)) * desiredPosition + (ctx.camRatio / (ctx.camRatio + 1.0f)) * precedingPosition;
+
+    if (gEnv->collisions && gEnv->collisions->forcecam)
+    {
+        gEnv->mainCamera->setPosition(gEnv->collisions->forcecampos);
+        gEnv->collisions->forcecam = false;
+    }
+    else
+    {
+        if (ctx.cct_player_actor && ctx.cct_player_actor->ar_replay_mode && camDisplacement != Vector3::ZERO)
+            gEnv->mainCamera->setPosition(desiredPosition);
+        else
+            gEnv->mainCamera->setPosition(camPosition);
+    }
+
+    ctx.camLookAtSmooth = (1.0f / (ctx.camRatio + 1.0f)) * ctx.camLookAt + (ctx.camRatio / (ctx.camRatio + 1.0f)) * precedingLookAt;
+
+    ctx.camLookAtLast = ctx.camLookAt;
+    ctx.camLookAtSmoothLast = ctx.camLookAtSmooth;
+    gEnv->mainCamera->lookAt(ctx.camLookAtSmooth);
+}
+
+bool CameraManager::CameraBehaviorOrbitMouseMoved( CameraManager::CameraContext& ctx, const OIS::MouseEvent& _arg)
+{
+    const OIS::MouseState ms = _arg.state;
+
+    if (ms.buttonDown(OIS::MB_Right))
+    {
+        float scale = RoR::App::GetInputEngine()->isKeyDown(OIS::KC_LMENU) ? 0.002f : 0.02f;
+        ctx.camRotX += Degree(ms.X.rel * 0.13f);
+        ctx.camRotY += Degree(-ms.Y.rel * 0.13f);
+        ctx.camDist += -ms.Z.rel * scale;
+        return true;
+    }
+
+    return false;
+}
+
+void CameraManager::CameraBehaviorOrbitReset( CameraManager::CameraContext& ctx)
+{
+    ctx.camRotX = 0.0f;
+    ctx.camRotXSwivel = 0.0f;
+    ctx.camRotY = 0.3f;
+    ctx.camRotYSwivel = 0.0f;
+    ctx.camLookAtLast = Vector3::ZERO;
+    ctx.camLookAtSmooth = Vector3::ZERO;
+    ctx.camLookAtSmoothLast = Vector3::ZERO;
+    gEnv->mainCamera->setFOVy(ctx.cct_fov_exterior);
+}
+
+void CameraManager::CameraBehaviorOrbitNotifyContextChange( CameraManager::CameraContext& ctx)
+{
+    ctx.camLookAtLast = Vector3::ZERO;
 }
